Move registerSmartSemicolonDetection to standardLanguageClient
Contributed by Roland Grunberg <rgrunber@redhat.com>
