+++ package.nls.json
@@ -25,6 +25,6 @@
 	"java.project.createModuleInfo.command": "Create module-info.java",
 	"java.clean.sharedIndexes": "Clean Shared Indexes",
 	"java.server.restart": "Restart Java Language Server",
-	"java.action.smartSemicolonDetection": "Java Smart Semicolon Detection",
-	"java.edit.smartSemiColonDetectionUndo": "Java Smart Semicolon Detection Undo"
+	"java.edit.smartSemicolonDetection": "Java Smart Semicolon Detection",
+	"java.edit.smartSemicolonDetectionUndo": "Java Smart Semicolon Detection Undo"
 }
